Refactor mappers to use Mappie and fix Kotlin hierarchy warnings
- Convert CurrencyMapper to use Mappie ObjectMappie pattern
- Add TransferMapper, AccountBalanceMapper, and AccountRowMapper using Mappie
- Refactor TransactionRepositoryImpl to use the new Mappie mappers
- Change Currency.scaleFactor from Int to Long for consistency with SQLite INTEGER
- Fix Kotlin multiplatform hierarchy warnings by adding gradle.properties
  to disable default hierarchy template in utils/bigdecimal and utils/currency
- Add kotlin-multiplatform-convention to both utils modules for expect/actual support

+++ app/db/core/src/commonMain/kotlin/com/moneymanager/database/repository/CategoryRepositoryImpl.kt
@@ -41,7 +41,7 @@ class CategoryRepositoryImpl(
                             id = CurrencyId(Uuid.parse(row.currency_id)),
                             code = row.currency_code,
                             name = row.currency_name,
-                            scaleFactor = row.currency_scaleFactor.toInt(),
+                            scaleFactor = row.currency_scaleFactor,
                         )
                     CategoryBalance(
                         categoryId = row.categoryId,
